cleanup useless codes. format,
