+++ tardis/src/cluster.rs
@@ -1,8 +1,8 @@
+pub mod cluster_broadcast;
+pub mod cluster_hashmap;
 pub mod cluster_processor;
 pub mod cluster_publish;
 pub mod cluster_receive;
 mod cluster_watch_by_cache;
 #[cfg(feature = "k8s")]
 mod cluster_watch_by_k8s;
-pub mod cluster_broadcast;
-pub mod cluster_hashmap;
