+++ tardis/src/cluster/cluster_watch_by_cache.rs
@@ -1,4 +1,4 @@
-use std::{net::SocketAddr, time::Duration, collections::HashSet};
+use std::{collections::HashSet, net::SocketAddr, time::Duration};
 
 use chrono::Utc;
 use tokio::time;
