+++ tardis/src/web/ws_processor/cluster_protocol.rs
@@ -1,23 +1,16 @@
-//!
-//!
-//!
-//!
-//! Protocol:
-//!
-//! 1. Route
-//! 2. Avatar
-//! 3. Forward
-
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
-use std::{borrow::Cow, collections::HashMap};
+use std::{borrow::Cow, collections::HashMap, sync::Arc};
 
 use crate::{
     basic::result::TardisResult,
-    cluster::cluster_processor::{TardisClusterMessageReq, TardisClusterSubscriber},
+    cluster::{
+        cluster_broadcast::ClusterBroadcastChannel,
+        cluster_processor::{TardisClusterMessageReq, TardisClusterSubscriber},
+    },
 };
 
-use super::ws_insts_mapping_avatars;
+use super::{ws_insts_mapping_avatars, TardisWebsocketMgrMessage, WsBroadcastSender};
 
 pub const EVENT_AVATAR: &str = "tardis/avatar";
 
@@ -26,7 +19,6 @@ pub(crate) struct Avatar;
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub(crate) enum AvatarMessage {
     Sync { table: HashMap<String, Vec<String>> },
-
 }
 
 #[async_trait::async_trait]
@@ -51,31 +43,22 @@ impl TardisClusterSubscriber for Avatar {
     }
 }
 
-// pub(crate) struct Forward {}
+impl WsBroadcastSender for ClusterBroadcastChannel<TardisWebsocketMgrMessage> {
+    fn subscribe(&self) -> tokio::sync::broadcast::Receiver<TardisWebsocketMgrMessage> {
+        self.local_broadcast_channel.subscribe()
+    }
 
-// #[derive(Debug, Clone, Serialize, Deserialize)]
-// pub(crate) enum ForwardMessage {
-//     Forward {
-//         to_inst: String,
-//         message: String,
-//     },
-// }
+    fn send(&self, msg: TardisWebsocketMgrMessage) {
+        self.send(msg);
+    }
+}
 
-// #[async_trait::async_trait]
-// impl TardisClusterSubscriber for Forward {
-//     fn event_name(&self) -> Cow<'static, str> {
-//         "cluster/forward".into()
-//     }
+impl WsBroadcastSender for Arc<ClusterBroadcastChannel<TardisWebsocketMgrMessage>> {
+    fn subscribe(&self) -> tokio::sync::broadcast::Receiver<TardisWebsocketMgrMessage> {
+        self.local_broadcast_channel.subscribe()
+    }
 
-//     async fn subscribe(&self, message_req: TardisClusterMessageReq) -> TardisResult<Option<Value>> {
-//         // let from_node = message_req.req_node_id;
-//         if let Ok(message) = serde_json::from_value(message_req.msg) {
-//             match message {
-//                 ForwardMessage::Forward { to_inst, message } => {
-
-//                 }
-//             }
-//         }
-//         Ok(None)
-//     }
-// }
+    fn send(&self, msg: TardisWebsocketMgrMessage) {
+        ClusterBroadcastChannel::send(self, msg);
+    }
+}
