+++ tardis/tests/test_cluster.rs
@@ -1,7 +1,6 @@
 use std::{
     borrow::Cow,
     env,
-    ffi::OsStr,
     path::Path,
     sync::atomic::{AtomicUsize, Ordering},
     time::Duration,
